SCB-85 move file output code from metrics-extension to samples
