--- metrics/metrics-sample/metrics-performance-test/src/main/resources/microservice.yaml
@@ -1,52 +1,0 @@
-## ---------------------------------------------------------------------------
-## Licensed to the Apache Software Foundation (ASF) under one or more
-## contributor license agreements.  See the NOTICE file distributed with
-## this work for additional information regarding copyright ownership.
-## The ASF licenses this file to You under the Apache License, Version 2.0
-## (the "License"); you may not use this file except in compliance with
-## the License.  You may obtain a copy of the License at
-##
-##      http://www.apache.org/licenses/LICENSE-2.0
-##
-## Unless required by applicable law or agreed to in writing, software
-## distributed under the License is distributed on an "AS IS" BASIS,
-## WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-## See the License for the specific language governing permissions and
-## limitations under the License.
-## ---------------------------------------------------------------------------
-
-# all interconnected microservices must belong to an application wth the same ID
-APPLICATION_ID: metricsPerfTest
-service_description:
-# name of the declaring microservice
-  name: demo
-  version: 0.0.1
-cse:
-  service:
-    registry:
-      address: http://127.0.0.1:30100
-  rest:
-    address: 0.0.0.0:7777
-
-servicecomb:
-  metrics:
-    #polltime,unit is second
-    polltime: 5
-    #metrics double value round places,default value is 1
-    round_places: 1
-    file:
-      enabled: false
-      #metrics file name prefix setting,default value is "metrics"
-      name_prefix: bmi.calculator
-
-      root_path: /sample/bmi/calculator/target/metric
-      rolling:
-        max_file_count: 10
-        max_file_size : 10MB
-
-    #Version 0.5.0+ config
-    #Support Muti-WindowTime (MILLISECONDS) like 10000,60000 (10s and 60s), Use WindowTimeIndex make selection,start from 0
-    #Default value is single WindowTime : 5000 (5 second)
-    #Will direct set value of "servo.pollers"
-    #More info can be found here : https://github.com/Netflix/servo/wiki/Getting-Started -> Polling Intervals Chapter
-    window_time: 5000
