+++ samples/metrics-write-file-sample/metrics-write-file-config/pom.xml
@@ -19,20 +19,14 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <packaging>pom</packaging>
   <parent>
-    <groupId>io.servicecomb</groupId>
-    <artifactId>java-chassis-parent</artifactId>
+    <artifactId>metrics-write-file-sample</artifactId>
+    <groupId>io.servicecomb.samples</groupId>
     <version>0.6.0-SNAPSHOT</version>
-    <relativePath>../parent</relativePath>
   </parent>
+  <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>metrics</artifactId>
+  <artifactId>metrics-write-file-config</artifactId>
 
-  <modules>
-    <module>metrics-common</module>
-    <module>metrics-core</module>
-    <module>metrics-extension</module>
-  </modules>
+
 </project>
